requirements.txt
openai>=1.0.0
feedparser>=6.0.0
python-dotenv>=1.0.0
requests>=2.28.0
slack-sdk>=3.19.5
beautifulsoup4>=4.11.1
schedule>=1.1.0
lxml>=4.9.0
html5lib>=1.1